# apps/api/contract/openapi.yaml
openapi: 3.1.0
info:
  title: LearnR API
  version: 0.1.0
...
paths:
...
  /shares:
    get:
      operationId: readShares
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        token:
                          type: string
                        childIds:
                          type: array
                          items:
                            type: string
                        createdAt:
                          type: string
                          format: date-time
                        expiresAt:
                          type: string
                          format: date-time
                      required:
                        - id
                        - token
                        - childIds
                        - createdAt
                        - expiresAt
                      additionalProperties: false
                  viewers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          anyOf:
                            - type: string
                            - type: 'null'
                        email:
                          anyOf:
                            - type: string
                            - type: 'null'
                        image:
                          anyOf:
                            - type: string
                            - type: 'null'
                        children:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              name:
                                type: string
                            required:
                              - id
                              - name
                            additionalProperties: false
                      required:
                        - id
                        - name
                        - email
                        - image
                        - children
                      additionalProperties: false
                required:
                  - invites
                  - viewers
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
    post:
      operationId: createShareInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                childIds:
                  minItems: 1
                  type: array
                  items:
                    type: string
              required:
                - childIds
      responses:
        '201':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                required:
                  - token
                  - expiresAt
                additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /shares/{token}:
    get:
      operationId: readShareInvite
      parameters:
        - schema:
            type: string
          in: path
          name: token
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ownerId:
                    type: string
                  ownerName:
                    anyOf:
                      - type: string
                      - type: 'null'
                  children:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        avatar:
                          type: string
                          enum:
                            - fox
                            - bear
                            - cat
                            - owl
                            - frog
                            - whale
                            - rabbit
                            - panda
                        photo:
                          anyOf:
                            - type: string
                            - type: 'null'
                        level:
                          anyOf:
                            - type: string
                            - type: 'null'
                      required:
                        - id
                        - name
                        - avatar
                        - photo
                        - level
                      additionalProperties: false
                  expiresAt:
                    type: string
                    format: date-time
                  live:
                    type: boolean
                required:
                  - ownerId
                  - ownerName
                  - children
                  - expiresAt
                  - live
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /shares/{id}:
    delete:
      operationId: cancelShareInvite
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '204':
          description: Default Response
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /shares/{token}/accept:
    post:
      operationId: acceptShareInvite
      parameters:
        - schema:
            type: string
          in: path
          name: token
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      ok:
                        type: boolean
                        enum:
                          - true
                      children:
                        type: integer
                        minimum: -9007199254740991
                        maximum: 9007199254740991
                    required:
                      - ok
                      - children
                    additionalProperties: false
                  - type: object
                    properties:
                      ok:
                        type: boolean
                        enum:
                          - false
                      reason:
                        type: string
                        enum:
                          - unavailable
                          - own-link
                          - error
                    required:
                      - ok
                      - reason
                    additionalProperties: false
  /shares/viewers/{viewerId}:
    delete:
      operationId: revokeShare
      parameters:
        - schema:
            type: string
          in: query
          name: childId
          required: false
        - schema:
            type: string
          in: path
          name: viewerId
          required: true
      responses:
        '204':
          description: Default Response
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /shares/mine/{childId}:
    delete:
      operationId: leaveShare
      parameters:
        - schema:
            type: string
          in: path
          name: childId
          required: true
      responses:
        '204':
          description: Default Response
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
...